iris: tracking region backing objects for initial kernel regions
In particular, the kernel heap and executable are now tracked using
proper backing objects. This is a step towards being able to lazily
map pages in the kernel heap, and to use this mechanism for all memory
management.

Additionally, di::StrongInt is changed to support directly adding
unsigned types, which prevents a lot of casting. Unfortunately, this
introduces an ambiguity when using integer literals, but this can be
fixed by providing the correct type explicitly.

+++ iris/arch/x86/amd64/hw/io_apic.cpp
@@ -80,7 +80,7 @@ void tag_invoke(di::Tag<enable_irq_line>, IoApic& self, IrqLine irq_line) {
 
     // Set the interrupt vector.
     redirection_entry.set<IoApicRedirectionTableEntryVector>(
-        (global_state().arch_readonly_state.external_irq_offset + i8(irq_line.raw_value())).raw_value());
+        (global_state().arch_readonly_state.external_irq_offset + irq_line.raw_value()).raw_value());
 
     // Set the delivery mode to fixed.
     redirection_entry.set<IoApicRedirectionTableEntryDeliveryMode>(ApicMessageType::Fixed);
